+++ src/main/java/com/buuz135/industrial/proxy/event/MobDeathHandler.java
@@ -1,5 +1,6 @@
-package com.buuz135.industrial.proxy;
+package com.buuz135.industrial.proxy.event;
 
+import com.buuz135.industrial.proxy.CommonProxy;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;